requirements.txt
Flask==3.0.3
num2words==0.5.13
text2digits==0.1.0
python-dotenv==1.0.0
requests==2.31.0
